+++ changelog.md
@@ -1,9 +1,10 @@
 ## 3.0.0
-
+- [FEA][COR] Toaster handler for add and like recipes
 - [FEA][COR] Singleton catalog filter service
 - [FEA][AND] Refact catalog header
 - [FEA][AND] My meal counter button
 - [FEA][IOS] My meal counter button
+- [FEA][COR] Picture is avalaible in recipes ingredients
 
 ## 2.5.2
 - [FIX][IOS] Pass browse catalog action to favorites empty view template